+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
+## [101.2.25] - 2026-09-09
+### Changed
+- Delegated the allAttributesLoaded check to a dedicated hasAllAttributeLoaded()
+
 ## [101.2.24] - 2026-08-21
 ### Added
 - "Reset Is Default" button in attribute options